% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{EucDistance}
\alias{EucDistance}
\title{EucDistance}
\usage{
EucDistance(x1, x2)
}
\arguments{
\item{x1}{corresponde ao um vetor numérico de comprimento n}

\item{x2}{corresponde ao um vetor numérico de comprimento n}
}
\value{
\code{EucDistance} retorna um inteiro correspondendo a distância entre
as duas séries passadas
}
\description{
Essa função retorna a distância euclidiana entre dois vetores de mesmo comprimento
}
\examples{
ts.1 <- runif(n = 100)
ts.2 <- runif(n = 100)
EucDistance(ts.1,ts.2)
}
