% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{SlidingWindow}
\alias{SlidingWindow}
\title{SlidingWindow}
\usage{
SlidingWindow(ts, n)
}
\arguments{
\item{ts}{corresponde a uma série temporal numérica}

\item{n}{corresponde ao tamanho da subsequência desejada}
}
\value{
\code{SlidingWindow} retorna um data.frame contendo todas a subsequências
de tamanho \code{n}
}
\description{
Esta função aplica a janela deslizante sobre uma série temporal
}
\examples{
ts <- runif(n = 100)
SlidingWindow(ts,10)
}
